+++ RELEASE_NOTES.md
@@ -33,4 +33,5 @@
 ## Version 1.6.5
 - Remove atomic operations on non atomic bool in scheduler
 - Support removed for Debian 11
+- Added function `iot_data_vector_to_list` to populate a list from a vector